+++ client/src/components/CardEvent.jsx
@@ -1,4 +1,6 @@
 import PropTypes from 'prop-types';
+import { format } from 'date-fns';
+import { fr } from 'date-fns/locale';
 
 import { HiMiniCalendarDays } from 'react-icons/hi2';
 import { TbClockHour2Filled } from 'react-icons/tb';
@@ -21,13 +23,13 @@ export default function CardEvent({ event }) {
         />
       </div>
       <div className="p-4 sm:pr-6 sm:pl-4">
+        <p className="block mb-2 font-sans text-xl antialiased min[1900px]:text-4xl 2xl:text-md font-semibold leading-snug tracking-normal normal-case transition-colors text-blue-gray-900">
+          {event.title}
+        </p>
         <p className="block mb-4 font-sans text-sm antialiased font-semibold leading-normal 2xl:text-lg text-sky-800">
           {event.subtitle}
         </p>
 
-        <p className="block mb-2 font-sans text-xl antialiased min[1900px]:text-4xl 2xl:text-md font-semibold leading-snug tracking-normal normal-case transition-colors text-blue-gray-900">
-          {event.title}
-        </p>
         <p className="block antialiased font-sans text-base min[1900px]:text-xl 2xl:text-md leading-relaxed 2xl:mb-4 font-normal !text-gray-500">
           {event.description}
         </p>
@@ -37,15 +39,19 @@ export default function CardEvent({ event }) {
             <TbClockHour2Filled />
           </div>
           <div className="flex flex-col">
-            <p className="antialiased text-base 2xl:text-xl text-blue-gray-900 mb-0.5 font-semibold">
-              <time dateTime={event.date}>{event.date}</time>
+            <p className="antialiased text-base capitalize 2xl:text-xl text-blue-gray-900 mb-0.5 font-semibold">
+              <time dateTime={event.date}>
+                {event.parsedDate
+                  ? format(event.parsedDate, 'dd MMMM yyyy', { locale: fr })
+                  : event.date}
+              </time>
             </p>
             <p className="mx-0 text-sm antialiased font-normal text-gray-700 2xl:text-md">
               {event.author.schedules}
             </p>
           </div>
-          <div className="flex flex-row ml-16 sm:ml-4 w-14">
-            <p className="px-2 py-1 mx-0 text-xs antialiased font-semibold border-none text-zinc-100 2xl:text-md rounded-xl ring-offset-sky-600 ring-2 bg-sky-600">
+          <div className="flex flex-row  align-items ml-16 md:ml-4 w-14">
+            <p className="px-2 py-1 mx-0 lining-nums text-base antialiased font-semibold border-none text-zinc-100 2xl:text-md rounded-xl ring-offset-sky-600 ring-2 bg-sky-600">
               {event.author.price}
             </p>
           </div>
@@ -64,6 +70,7 @@ CardEvent.propTypes = {
     description: PropTypes.string.isRequired,
     subtitle: PropTypes.string.isRequired,
     date: PropTypes.string.isRequired,
+    parsedDate: PropTypes.number.isRequired,
     author: PropTypes.shape({
       name: PropTypes.string.isRequired,
       schedules: PropTypes.string.isRequired,